Bug: Fix decimal resolution.

+++ PinnacleFeed/Pinnacle.Feed.Database/Schema Objects/Schemas/dbo/Tables/Spread.table.sql
@@ -1,11 +1,11 @@
 ﻿CREATE TABLE [dbo].[Spread]
 (
-	MatchId		bigint		NOT NULL, 
-	SportId		bigint		NOT NULL, 
-	LeagueId	bigint		NOT NULL, 
-	HomeSpread  decimal		NOT NULL, 
-	AwaySpread  decimal		NOT NULL,
-	HomePrice	decimal		NOT NULL, 
-	AwayPrice	decimal		NOT NULL,
-	IsAlt		bit			NOT NULL
+	MatchId		bigint			NOT NULL, 
+	SportId		bigint			NOT NULL, 
+	LeagueId	bigint			NOT NULL, 
+	HomeSpread  decimal(3,2)	NOT NULL, 
+	AwaySpread  decimal(3,2)	NOT NULL,
+	HomePrice	decimal(3,2)	NOT NULL, 
+	AwayPrice	decimal(3,2)	NOT NULL,
+	IsAlt		bit				NOT NULL
 )
