+++ PinnacleFeed/Pinnacle.Feed.Database/Schema Objects/Schemas/dbo/Tables/Total.table.sql
@@ -1,10 +1,10 @@
 ﻿CREATE TABLE [dbo].[Total]
 (
-	MatchId		bigint		NOT NULL, 
-	SportId		bigint		NOT NULL, 
-	LeagueId	bigint		NOT NULL, 
-	Points		decimal		NOT NULL, 	
-	OverPrice	decimal		NOT NULL, 
-	UnderPrice	decimal		NOT NULL,
-	IsAlt		bit			NOT NULL
+	MatchId		bigint			NOT NULL, 
+	SportId		bigint			NOT NULL, 
+	LeagueId	bigint			NOT NULL, 
+	Points		decimal(3,2)	NOT NULL, 	
+	OverPrice	decimal(3,2)	NOT NULL, 
+	UnderPrice	decimal(3,2)	NOT NULL,
+	IsAlt		bit				NOT NULL
 )
